fix: update project directory
