--- src/main/java/com/hyundai/config/ApplicationProperties.java
@@ -1,15 +1,0 @@
-package com.hyundai.config;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ApplicationProperties {
-
-    private String FRONT_LOCAL_URL;
-    private String DOMAIN;
-    private String KAKAO_REDIRECT_URI;
-    private String KAKAO_CLIENT_ID;
-    private String JWT_SECRET_KEY;
-}
