+++ src/main/java/com/hyundai/domain/login/controller/AdminController.java
@@ -1,0 +1,15 @@
+package com.hyundai.domain.login.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequestMapping("/api")
+@RestController
+public class AdminController {
+
+    @RequestMapping("/test")
+    public ResponseEntity<String> test(){
+        return ResponseEntity.ok().body("WillEat api ok");
+    }
+}
