+++ src/main/java/com/hyundai/domain/login/dto/Criteria.java
@@ -1,0 +1,17 @@
+package com.hyundai.domain.login.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+@Getter
+@SuperBuilder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Criteria {
+    private String selectId;
+    private int pageNum;
+    private int pageSize;
+    private String sorted;
+}
