+++ src/main/java/com/hyundai/domain/login/dto/MemberListDto.java
@@ -1,0 +1,14 @@
+package com.hyundai.domain.login.dto;
+
+import lombok.*;
+
+import java.util.List;
+
+@Builder
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+@ToString
+public class MemberListDto {
+    List<MemberDto> memberList;
+}
