+++ src/main/java/com/hyundai/domain/login/entity/enumtype/OauthProvider.java
@@ -1,0 +1,6 @@
+package com.hyundai.domain.login.entity.enumtype;
+
+public enum OauthProvider {
+    KAKAO;
+
+}
