+++ src/main/java/com/hyundai/domain/login/entity/enumtype/Role.java
@@ -1,0 +1,5 @@
+package com.hyundai.domain.login.entity.enumtype;
+
+public enum Role {
+    ROLE_USER, ROLE_GUEST ,ROLE_ADMIN;
+}
