+++ src/main/java/com/hyundai/domain/login/security/dto/LoginRequestDto.java
@@ -1,0 +1,16 @@
+package com.hyundai.domain.login.security.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class LoginRequestDto {
+
+    private String token;
+    private String oauthType;
+}
