+++ src/main/java/com/hyundai/domain/login/security/dto/LoginResponseDto.java
@@ -1,0 +1,17 @@
+package com.hyundai.domain.login.security.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class LoginResponseDto {
+
+    private Boolean isFirstLogin;
+    private String accessToken;
+    private String refreshToken;
+}
