+++ src/main/java/com/hyundai/domain/login/security/dto/OAuthType.java
@@ -1,0 +1,10 @@
+package com.hyundai.domain.login.security.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum OAuthType {
+    KAKAO;
+}
