+++ src/main/java/com/hyundai/domain/login/security/dto/TokenDto.java
@@ -1,0 +1,16 @@
+package com.hyundai.domain.login.security.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class TokenDto {
+
+    private String accessToken;
+    private String refreshToken;
+}
