+++ src/main/java/com/hyundai/global/config/PropertyConfig.java
@@ -1,0 +1,18 @@
+package com.hyundai.global.config;
+
+import org.springframework.beans.factory.config.PropertiesFactoryBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.io.ClassPathResource;
+
+@Configuration
+public class PropertyConfig {
+    @Bean(name = "systemProperties")
+    public PropertiesFactoryBean propertiesFactoryBean() throws Exception {
+        PropertiesFactoryBean propertiesFactoryBean = new PropertiesFactoryBean();
+        ClassPathResource classPathResource = new ClassPathResource("config.properties");
+        propertiesFactoryBean.setLocation(classPathResource);
+        return propertiesFactoryBean;
+    }
+
+}
