--- src/main/java/com/hyundai/login/domain/enumtype/OauthProvider.java
@@ -1,6 +1,0 @@
-package com.hyundai.login.domain.enumtype;
-
-public enum OauthProvider {
-    KAKAO;
-
-}
