--- src/main/java/com/hyundai/login/domain/enumtype/Role.java
@@ -1,5 +1,0 @@
-package com.hyundai.login.domain.enumtype;
-
-public enum Role {
-    ROLE_USER, ROLE_GUEST ,ROLE_ADMIN;
-}
