--- src/main/java/com/hyundai/login/dto/MemberListDto.java
@@ -1,14 +1,0 @@
-package com.hyundai.login.dto;
-
-import lombok.*;
-
-import java.util.List;
-
-@Builder
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@ToString
-public class MemberListDto {
-    List<MemberDto> memberList;
-}
