--- src/main/java/com/hyundai/login/security/JwtResolver.java
@@ -1,23 +1,0 @@
-package com.hyundai.login.security;
-
-import io.jsonwebtoken.Claims;
-import lombok.extern.log4j.Log4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
-
-@Log4j
-@Component
-public class JwtResolver {
-
-    @Autowired
-    private JwtTokenProvider jwtTokenProvider;
-
-    public String getMemberIdByAccessToken(String accessToken) {
-        String token = jwtTokenProvider.resolveBearerToken(accessToken);
-        Claims claims = jwtTokenProvider.getClaims(token);
-        String memberId = claims.get("memberId").toString();
-        log.info("claims : " + claims + " memberId : " + memberId);
-
-        return memberId;
-    }
-}
