--- src/main/java/com/hyundai/login/security/dto/KakaoOauthTokenDto.java
@@ -1,28 +1,0 @@
-package com.hyundai.login.security.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class KakaoOauthTokenDto {
-
-    @JsonProperty("access_token")
-    private String accessToken;
-
-    @JsonProperty("token_type")
-    private String tokenType;
-
-    @JsonProperty("refresh_token")
-    private String refreshToken;
-
-    @JsonProperty("expires_in")
-    private String expiresIn;
-
-    @JsonProperty("refresh_token_expires_in")
-    private String refreshTokenExpiresIn;
-
-    @JsonProperty("scope")
-    private String scope;
-}
