--- src/main/java/com/hyundai/login/security/dto/KakaoUserInfoDto.java
@@ -1,30 +1,0 @@
-package com.hyundai.login.security.dto;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-
-@Getter
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class KakaoUserInfoDto {
-
-    @JsonProperty("kakao_account")
-    private KakaoAccount kakaoAccount;
-
-    @Getter
-    @JsonIgnoreProperties(ignoreUnknown = true)
-    static class KakaoAccount {
-        private KakaoProfile profile;
-        private String email;
-    }
-
-    @Getter
-    @JsonIgnoreProperties(ignoreUnknown = true)
-    static class KakaoProfile {
-        private String nickname;
-    }
-
-    public String getEmail() {
-        return kakaoAccount.email;
-    }
-}
