--- src/main/java/com/hyundai/login/security/dto/LoginRequestDto.java
@@ -1,16 +1,0 @@
-package com.hyundai.login.security.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class LoginRequestDto {
-
-    private String token;
-    private String oauthType;
-}
