--- src/main/java/com/hyundai/login/security/dto/OAuthType.java
@@ -1,10 +1,0 @@
-package com.hyundai.login.security.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum OAuthType {
-    KAKAO;
-}
