+++ src/main/resources/com/hyundai/mapper/MemberMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.hyundai.login.mapper.MemberMapper">
+<mapper namespace="com.hyundai.domain.login.mapper.MemberMapper">
 
     <select id="findMemberByEmail" resultType="Member">
         SELECT id
