+++ src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
@@ -28,7 +28,7 @@
 	<interceptors>
 		<interceptor>
 			<mapping path="/api/v1/auth/**"/>
-			<beans:bean id="jwtInterceptor" class="com.hyundai.login.security.JwtInterceptor" />
+			<beans:bean id="jwtInterceptor" class="com.hyundai.domain.login.security.JwtInterceptor" />
 		</interceptor>
 	</interceptors>
